+++ src/Fable.Transforms/Rust/Fable2Rust.fs
@@ -300,22 +300,6 @@ module UsageTracking =
 
 module TypeInfo =
 
-    let splitName (sep: string) (fullName: string) =
-        let i = fullName.LastIndexOf(sep)
-
-        if i < 0 then
-            "", fullName
-        else
-            fullName.Substring(0, i), fullName.Substring(i + sep.Length)
-
-    let splitLast (fullName: string) =
-        let i = fullName.LastIndexOf(".")
-
-        if i < 0 then
-            fullName
-        else
-            fullName.Substring(i + 1)
-
     let makeFullNamePath fullName genArgsOpt =
         let parts = splitNameParts fullName
         mkGenericPath parts genArgsOpt
@@ -680,12 +664,12 @@ module TypeInfo =
                     makeFullNamePathExpr importName genArgsOpt
                 else
                     // for constructors or static members, import just the type
-                    let selector, membName = splitName "." info.Selector
+                    let selector, membName = Fable.Naming.splitLastBy "." info.Selector
                     let importName = com.GetImportName(ctx, selector, info.Path, r)
                     makeFullNamePathExpr (importName + "::" + membName) genArgsOpt
             | Fable.LibraryImport mi when not (mi.IsInstanceMember) && not (mi.IsModuleMember) ->
                 // for static (non-module and non-instance) members, import just the type
-                let selector, membName = splitName "::" info.Selector
+                let selector, membName = Fable.Naming.splitLastBy "::" info.Selector
                 let importName = com.GetImportName(ctx, selector, info.Path, r)
                 makeFullNamePathExpr (importName + "::" + membName) genArgsOpt
             | _ ->
@@ -1265,7 +1249,7 @@ module Util =
         let args =
             args |> discardUnitArg [] |> List.filter (fun arg -> not (arg.IsThisArgument))
 
-        let label = splitLast name
+        let label = Fable.Naming.splitLast name
 
         interface ITailCallOpportunity with
             member _.Label = label
@@ -2303,9 +2287,10 @@ module Util =
             expr |> mkParenExpr // if not an identifier, wrap it in parentheses
 
     let transformCall (com: IRustCompiler) ctx range (typ: Fable.Type) calleeExpr (callInfo: Fable.CallInfo) =
+        let membOpt = callInfo.MemberRef |> Option.bind com.TryGetMember
+
         let isByRefPreferred =
-            callInfo.MemberRef
-            |> Option.bind com.TryGetMember
+            membOpt
             |> Option.map (fun memberInfo ->
                 memberInfo.Attributes
                 |> Seq.exists (fun a -> a.Entity.FullName = Atts.rustByRef)
@@ -2313,8 +2298,7 @@ module Util =
             |> Option.defaultValue false
 
         let argParams =
-            callInfo.MemberRef
-            |> Option.bind com.TryGetMember
+            membOpt
             |> Option.map (fun memberInfo -> memberInfo.CurriedParameterGroups |> List.concat)
             |> Option.defaultValue []
 
@@ -2337,10 +2321,16 @@ module Util =
                 IsParamByRefPreferred = isByRefPreferred
             }
 
-        let args =
+        let callArgs =
             FSharp2Fable.Util.dropUnitCallArg callInfo.Args callInfo.SignatureArgTypes
 
-        let args = transformCallArgs com ctx args callInfo.SignatureArgTypes argParams
+        let args = transformCallArgs com ctx callArgs callInfo.SignatureArgTypes argParams
+
+        let genArgsOpt =
+            if List.isEmpty callArgs then
+                transformGenArgs com ctx callInfo.GenericArgs
+            else
+                None
 
         match calleeExpr with
         // mutable module values (transformed as function calls)
@@ -2362,28 +2352,7 @@ module Util =
             // library imports without args need explicit genArgs
             // this is for imports like Array.empty, Seq.empty etc.
             let needGenArgs =
-                Set.ofList
-                    [
-                        "Native_::defaultOf"
-                        "Native_::getZero"
-                        "NativeArray_::new_empty"
-                        "NativeArray_::new_with_capacity"
-                        "HashSet_::new_empty"
-                        "HashSet_::new_with_capacity"
-                        "HashMap_::new_empty"
-                        "HashMap_::new_with_capacity"
-                        "Set_::empty"
-                        "Map_::empty"
-                        "Seq_::empty"
-                    ]
-
-            let genArgsOpt =
-                if needGenArgs |> Set.contains info.Selector then
-                    match typ with
-                    | Fable.Tuple _ -> transformGenArgs com ctx [ typ ]
-                    | _ -> transformGenArgs com ctx typ.Generics // callInfo.GenericArgs
-                else
-                    None
+                List.isEmpty callArgs || info.Selector.EndsWith("::new_with_capacity")
 
             match callInfo.ThisArg, info.Kind with
             | Some thisArg, Fable.MemberImport memberRef ->
@@ -2393,17 +2362,25 @@ module Util =
                     let callee = transformCallee com ctx thisArg
                     mkMethodCallExpr info.Selector None callee args
                 else
-                    let callee = transformImport com ctx r t info None
+                    let callee = transformImport com ctx r t info genArgsOpt
                     mkCallExpr callee args
             | None, Fable.MemberImport memberRef when isModuleMemberRef com memberRef ->
                 let memb = com.GetMember(memberRef)
-                let callee = transformImport com ctx r t info None
+                let callee = transformImport com ctx r t info genArgsOpt
 
                 if memb.IsMutable && memb.IsValue then
                     mutableGet (mkCallExpr callee [])
                 else
                     mkCallExpr callee args
-            | None, Fable.LibraryImport _ ->
+            | None, Fable.LibraryImport memberInfo ->
+                let genArgsOpt =
+                    if needGenArgs && memberInfo.IsModuleMember then
+                        match typ with
+                        | Fable.Tuple _ -> transformGenArgs com ctx [ typ ]
+                        | _ -> transformGenArgs com ctx typ.Generics
+                    else
+                        None
+
                 let callee = transformImport com ctx r t info genArgsOpt
                 mkCallExpr callee args
             | _ ->
@@ -2419,9 +2396,9 @@ module Util =
                 | Some thisArg, Fable.IdentExpr ident ->
                     let callee = transformCallee com ctx thisArg
                     mkMethodCallExpr ident.Name None callee args
-                // | None, Fable.IdentExpr ident ->
-                //     let callee = makeFullNamePathExpr ident.Name None
-                //     mkCallExpr callee args
+                | None, Fable.IdentExpr ident ->
+                    let callee = makeFullNamePathExpr ident.Name genArgsOpt
+                    mkCallExpr callee args
                 | _ ->
                     let callee = transformCallee com ctx calleeExpr
                     mkCallExpr callee args
@@ -2435,7 +2412,7 @@ module Util =
         mkMethodCallExpr "set" None expr [ value ]
 
     let makeInstanceCall com ctx memberName calleeExpr args =
-        let membName = splitLast memberName
+        let membName = Fable.Naming.splitLast memberName
         let callee = com.TransformExpr(ctx, calleeExpr)
         // match calleeExpr.Type with
         // | IsNonErasedInterface com (entRef, genArgs) ->
@@ -3323,7 +3300,7 @@ module Util =
 
             memb.Attributes
             |> Seq.tryFind (fun att -> att.Entity.FullName = Atts.entryPoint)
-            |> Option.map (fun _ -> [ splitLast decl.Name ])
+            |> Option.map (fun _ -> [ Fable.Naming.splitLast decl.Name ])
         | Fable.ActionDeclaration decl -> None
         | Fable.ClassDeclaration decl -> None
 
@@ -3737,11 +3714,8 @@ module Util =
 
                     if ent.IsInterface then
                         let nameParts = getInterfaceImportName com ctx entRef |> splitNameParts
-
                         let genArgsOpt = transformGenArgs com ctx genArgs
-
                         let traitBound = mkTypeTraitGenericBound nameParts genArgsOpt
-
                         [ traitBound ]
                     else
                         []
@@ -3767,19 +3741,29 @@ module Util =
             mkLifetimeGenericBound "'static" //TODO: add it only when needed
         ]
 
-    let makeGenericParams com ctx (genArgs: Fable.Type list) =
+    let makeGenericParams com ctx (genParams: Fable.GenericParam list) =
+        genParams
+        |> List.filter (fun p -> not p.IsMeasure)
+        |> List.map (fun p ->
+            let typeBounds = makeTypeBounds com ctx p.Name p.Constraints
+            mkGenericParamFromName [] p.Name (typeBounds @ defaultTypeBounds)
+        )
+
+    let makeGenericParamsFromArgs com ctx (genArgs: Fable.Type list) =
         genArgs
         |> List.choose (
             function
             | Fable.GenericParam(name, isMeasure, constraints) when not isMeasure ->
                 let typeBounds = makeTypeBounds com ctx name constraints
-                let p = mkGenericParamFromName [] name (typeBounds @ defaultTypeBounds)
-                Some p
+                mkGenericParamFromName [] name (typeBounds @ defaultTypeBounds) |> Some
             | _ -> None
         )
 
     let makeGenerics com ctx (genArgs: Fable.Type list) =
-        makeGenericParams com ctx genArgs |> mkGenerics
+        makeGenericParamsFromArgs com ctx genArgs |> mkGenerics
+
+    let makeMemberGenerics com ctx (genParams: Fable.GenericParam list) =
+        makeGenericParams com ctx genParams |> mkGenerics
 
     let makeNestedFuncCtx com ctx (ident: Fable.Ident) usages =
         let scopedVarAttrs =
@@ -3929,7 +3913,7 @@ module Util =
         | _ -> []
 
     let transformModuleFunction (com: IRustCompiler) ctx (memb: Fable.MemberFunctionOrValue) (decl: Fable.MemberDecl) =
-        let name = splitLast decl.Name
+        let name = Fable.Naming.splitLast decl.Name
         //if name = "someProblematicFunction" then System.Diagnostics.Debugger.Break()
         let isByRefPreferred =
             memb.Attributes |> Seq.exists (fun a -> a.Entity.FullName = Atts.rustByRef)
@@ -3947,7 +3931,7 @@ module Util =
                 mkExprBlock fnBody
 
         let header = makeFnHeader com ctx memb.Attributes
-        let generics = makeGenerics com ctx genArgs
+        let generics = makeMemberGenerics com ctx memb.GenericParameters
         let kind = mkFnKind header fnDecl generics (Some fnBodyBlock)
         let attrs = transformAttributes com ctx memb.Attributes memb.XmlDoc
         let fnItem = mkFnItem attrs name kind
@@ -3961,7 +3945,7 @@ module Util =
         //     static value: OnceInit<T> = OnceInit::new();
         //     value.get_or_init(|| initValue).clone()
         // }
-        let name = splitLast decl.Name
+        let name = Fable.Naming.splitLast decl.Name
         let typ = decl.Body.Type
 
         let initNone = makeNew com ctx "Native" "OnceInit" []
@@ -4061,7 +4045,7 @@ module Util =
         (body: Fable.Expr)
         =
         let ctx = { ctx with IsAssocMember = true }
-        let name = splitLast membName
+        let name = Fable.Naming.splitLast membName
 
         let body =
             if memb.IsInstance && not (memb.IsConstructor) then
@@ -4138,7 +4122,7 @@ module Util =
 
     let transformAbbrev (com: IRustCompiler) ctx (ent: Fable.Entity) (decl: Fable.ClassDecl) =
         // TODO: this is unfinished and untested
-        let entName = splitLast ent.FullName
+        let entName = Fable.Naming.splitLast ent.FullName
         let genArgs = FSharp2Fable.Util.getEntityGenArgs ent
         let genArgsOpt = transformGenArgs com ctx genArgs
         let traitBound = mkTypeTraitGenericBound [ entName ] genArgsOpt
@@ -4150,7 +4134,7 @@ module Util =
         [ tyItem ]
 
     let transformUnion (com: IRustCompiler) ctx (ent: Fable.Entity) (decl: Fable.ClassDecl) =
-        let entName = splitLast ent.FullName
+        let entName = Fable.Naming.splitLast ent.FullName
         let genArgs = FSharp2Fable.Util.getEntityGenArgs ent
         let generics = makeGenerics com ctx genArgs
 
@@ -4181,7 +4165,7 @@ module Util =
         enumItem
 
     let transformClass (com: IRustCompiler) ctx (ent: Fable.Entity) (decl: Fable.ClassDecl) =
-        let entName = splitLast ent.FullName
+        let entName = Fable.Naming.splitLast ent.FullName
         let genArgs = FSharp2Fable.Util.getEntityGenArgs ent
         let generics = makeGenerics com ctx genArgs
         let isPublic = ent.IsFSharpRecord
@@ -4276,11 +4260,9 @@ module Util =
                         function
                         | Fable.Set(Fable.Value(Fable.ThisValue _, _), Fable.SetKind.FieldSet(fieldName), t, value, r) ->
                             let identExpr = identMap |> Map.find fieldName |> Fable.IdentExpr
-
                             Fable.Set(identExpr, Fable.ValueSet, t, value, r)
                         | Fable.Get(Fable.Value(Fable.ThisValue _, _), Fable.GetKind.FieldGet info, t, r) ->
                             let identExpr = identMap |> Map.find info.Name |> Fable.IdentExpr
-
                             identExpr
                         | e -> e
                     )
@@ -4303,13 +4285,14 @@ module Util =
         let fnItem = fnItem |> memberAssocItemWithVis com ctx memb
         fnItem
 
-    let makeInterfaceItems (com: IRustCompiler) ctx hasBody (ent: Fable.Entity) =
+    let makeInterfaceItems (com: IRustCompiler) ctx hasBody typeName (ent: Fable.Entity) =
         ent.AllInterfaces
         |> Seq.collect (fun ifc ->
             let ifcTyp = Fable.DeclaredType(ifc.Entity, ifc.GenericArgs)
             let ifcEnt = com.GetEntity(ifc.Entity)
 
             ifcEnt.MembersFunctionsAndValues
+            |> Seq.filter (fun memb -> memb.IsDispatchSlot)
             |> Seq.distinctBy (fun memb -> memb.CompiledName)
             |> Seq.map (fun memb ->
                 let thisArg = { makeTypedIdent ifcTyp "this" with IsThisArgument = true }
@@ -4324,19 +4307,25 @@ module Util =
                         makeTypedIdent p.Type name
                     )
 
-                let args = (thisArg :: memberArgs)
+                let args =
+                    if memb.IsInstance then
+                        thisArg :: memberArgs
+                    else
+                        memberArgs
 
                 let bodyOpt =
                     if hasBody then
-                        // let thisExpr = makeThis com ctx None ifcTyp
-                        let thisExpr = mkGenericPathExpr [ rawIdent "self" ] None
-
-                        let callee = thisExpr |> mkDerefExpr |> mkDerefExpr
-
                         let args = memberArgs |> List.map (transformIdent com ctx None)
 
-                        let name = memb.CompiledName
-                        let body = mkMethodCallExpr name None callee args
+                        let body =
+                            if memb.IsInstance then
+                                // let thisExpr = makeThis com ctx None ifcTyp
+                                let thisExpr = mkGenericPathExpr [ rawIdent "self" ] None
+                                let callee = thisExpr |> mkDerefExpr |> mkDerefExpr
+                                mkMethodCallExpr membName None callee args
+                            else
+                                makeCall [ typeName; membName ] None args
+
                         [ mkExprStmt body ] |> mkBlock |> Some
                     else
                         None
@@ -4346,11 +4335,13 @@ module Util =
         )
 
     let transformInterface (com: IRustCompiler) ctx (ent: Fable.Entity) (decl: Fable.ClassDecl) =
-        let entName = splitLast ent.FullName
+        let entName = Fable.Naming.splitLast ent.FullName
         let genArgs = FSharp2Fable.Util.getEntityGenArgs ent
+        let genArgNames = getEntityGenParamNames ent
+        let typeName = makeUniqueName "V" genArgNames
 
         let traitItem =
-            let assocItems = makeInterfaceItems com ctx false ent
+            let assocItems = makeInterfaceItems com ctx false typeName ent
             let generics = makeGenerics com ctx genArgs
             // let sendBound = mkTypeTraitGenericBound [ rawIdent "Send" ] None
             // let syncBound = mkTypeTraitGenericBound [ rawIdent "Sync" ] None
@@ -4359,14 +4350,12 @@ module Util =
             mkTraitItem attrs entName assocItems traitBounds generics
 
         let implItem =
-            let memberItems = makeInterfaceItems com ctx true ent
-            let genArgNames = getEntityGenParamNames ent
-            let typeName = makeUniqueName "V" genArgNames
+            let memberItems = makeInterfaceItems com ctx true typeName ent
             let genArgsOpt = transformGenArgs com ctx genArgs
             let traitBound = mkTypeTraitGenericBound [ entName ] genArgsOpt
             let typeBounds = traitBound :: defaultTypeBounds
             let typeParam = mkGenericParamFromName [] typeName typeBounds
-            let genParams = makeGenericParams com ctx genArgs
+            let genParams = makeGenericParamsFromArgs com ctx genArgs
             let generics = typeParam :: genParams |> mkGenerics
             let ty = mkGenericTy [ typeName ] [] |> makeLrcPtrTy com ctx
             let path = mkGenericPath [ entName ] genArgsOpt
@@ -4383,7 +4372,7 @@ module Util =
         //     }
         // }
         if ent.IsFSharpExceptionDeclaration then
-            let entName = splitLast ent.FullName
+            let entName = Fable.Naming.splitLast ent.FullName
             let entNameExpr = Fable.Value(Fable.StringConstant(entName), None)
 
             let thisArg = Fable.Value(Fable.ThisValue Fable.Any, None)
@@ -4501,9 +4490,7 @@ module Util =
                 )
 
             let macroName = getLibraryImportName com ctx "Native" op_macro
-
             let id_tokens = [ op_trait; op_fn; decl.Name ] |> List.map mkIdentToken
-
             let ty_tokens = (self_ty :: rhs_tys) @ genArgTys |> List.map mkTyToken
 
             let implItem =
@@ -4588,7 +4575,7 @@ module Util =
                 classDecl.Name // for interface object expressions
             else
                 getEntityFullName com ctx entRef
-            |> splitLast
+            |> Fable.Naming.splitLast
 
         let entType = FSharp2Fable.Util.getEntityType ent
         let genArgs = FSharp2Fable.Util.getEntityGenArgs ent